requirements.txt
Flask
robotframework
robotframework-selenium2library
selenium==3.4.0